Add support for IPv6 in firewall task / template
Uses the same ghetto_ipsets that are used by the IPv4 template, but
with an added item in each rule's list for `ipv6src`.

Right now this is only for Ubuntu hosts (tested on 14.04 and 12.04),
and even though only our Linode hosts actually use IPv6, it doesn't
hurt to run these everywhere actually.

Signed-off-by: Alan Orth <a.orth@cgiar.org>