roles/common: Remove mailrc template/task
It was used for RAID monitoring but the SELinux policy in CentOS 6
blocks crond from reading root's .mailrc, so we had to circumvent
it by manually specifying the args to mail using -S anyways.

Using one this method allows us to do the same thing on both Ubuntu
and CentOS hosts.

Signed-off-by: Alan Orth <a.orth@cgiar.org>