Clean up play
Instead of using lsof & grep'ing for daemons that are using deleting
libssl.so library, we can just reload either nginx or apache after
checking whether either is installed.

We'll need to re-evaluate how we check for the existence web services
when we start using systemd on production. Currently, we're still using
upstart on production.

Signed-off-by: James Oguya <j.oguya@cgiar.org>