roles/common: Import sshd_config template for Ubuntu 16.04
Started from a clean sshd_config, and merged with relevant settings
from the Ubuntu 15 config. Cipher support is obviously better in
newer OpenSSH but these settings are still good. PermitRootLogin
is the only setting with a new default, as upstream added this new
option to strictly disallow password logins for root.

Signed-off-by: Alan Orth <alan.orth@gmail.com>