roles/munin: Separate munin installation tasks
Previously, we were using one task to install munin on both Debian &
RedHat based distros. At times this task would fail on Debian-based
distros if apt cache is not updated, so the easiest way to fix this is
to have 2 tasks; one for Debian & the other for RedHat based distros.

Signed-off-by: James Oguya <j.oguya@cgiar.org>