roles/dspace: Block robots from accessing dynamic pages in nginx vhost
We had been doing it in robots.txt, but it turns out that you can't
use "*" for URL patterns in robots.txt. The only option is to block
these URLs in the webserver via X-Robots-Tag.

A few URLs to test with:

- https://cgspace.cgiar.org/handle/10568/440/browse?type=bioversity
- https://cgspace.cgiar.org/handle/10568/913/discover
- https://cgspace.cgiar.org/handle/10568/1/search-filter?filtertype_0=country&filter_0=VIETNAM&filter_relational_operator_0=equals&field=country

Test with `httpie --print Hh` to see the header.