roles/compute: Separate service tasks for CentOS 6 & 7
In CentOS 6, we were using a wrapper init.d script which would determine
if a node is a head node or just a normal compute node. In CentOS 7, the
same init.d script doesn't work, so we'll have to use systemd service
files to achieve the same effect. The only difference is starting
slurmctld service on the head node & slurmd service on other compute
nodes.

Signed-off-by: James Oguya <j.oguya@cgiar.org>