roles/compute: Add two separate tasks for copying slurm.conf to CentOS 6 & 7
Slurm on CentOS 6 can be restarted by using the wrapper init.d slurm
script but not on CentOS 7. You have to use either slurmd, slurmctld or
both systemd service files to achieve the same effect.
Slurm on both versions uses the same directory, but the only problem
comes in when firing a handler task & differentiating between the two
versions. Therefore, the easiest solution is to have two separate tasks.

Signed-off-by: James Oguya <j.oguya@cgiar.org>