roles/common: Fix bug in backup_mysql.py script
Providing a password to mysql's -p option raises a warning on MySQL 5.7
& this ends up making the script to erroneously backup missing databases.
Besides that, this is insecure because the password can read from `ps'
or `/proc' fs. To fix this, this script will write auth(user/pass) info to
a temp file which is then used with the `--default-extra-file' option on
both mysql & mysqldump cli tools.

Fixes #36

Signed-off-by: James Oguya <j.oguya@cgiar.org>