roles/dspace: Install Java meta package
Instead of installing the Java "installer" package from the WebUpd8
PPA, we should be using the meta package that installs the version
we want as well as setting it to be the system default.

We essentially get this by installing the "installer" package, but
if there is any possibility of a host having two versions of Java,
for example due to host "drift" from the playbooks, we should be
more idempotent and install the "set default" package instead, as
it depends on the "installer" package anyways.