roles/compute: Add compute03 to slurm.conf
compute03 was bought by BecA & ICT back in 2014 for hosting some JIC
websites which we migrated to Linode. See commit
b61f462e067f893c129efbd838295dd4116ccc25 on
github.com/ilri/rmg-ansible-private for more details on this.

Am particularly interested in this server because it's the only server
we have that could use the remaining 32GB RAM sticks we bought for
mammoth :). It now has 442GB RAM which is good for highmem jobs.

Ideally, I should be cherrypicking this from commit
4728e31cca8fa2b0f1b9d35f4246158e8f2970d8 in centos7 branch.

Signed-off-by: James Oguya <j.oguya@cgiar.org>