roles/dspace: Use proxy_pass to upstream in nginx
Using proxy_pass to an upstream server (instead of directly to the
server's IP) allows us to let nginx use TCP keepalive when talking
to the backend server as well as simplify the config a bit.

As it is now, nginx creates a new TCP connection to the backend for
each client that connects to the front end—this enables it to keep
some of the connections open so it can re-use them.

See: https://www.nginx.com/blog/http-keepalives-and-web-performance/
See: https://www.nginx.com/blog/tuning-nginx/
See: https://nginx.org/en/docs/http/ngx_http_upstream_module.html#keepalive