roles/dspace: Fix cron job
One scheduled task on the DSpace documentation wiki was incorrectly
running on every day instead of just Sunday. We don't use this one
but it's good to keep the task list up to date so that there are
less changes to look over when we compare to the official ones.

See: https://wiki.duraspace.org/display/DSDOC5x/Scheduled+Tasks+via+Cron