roles/dspace: nginx redirects for CGIAR Library
There are many links to the CGIAR Library's browse pages on various
websites so we need to redirect them where possible. Mostly we just
send their top-level queries to their community, but in the case of
the "subject" browse we need to send those to a new "systemsubject"
browse index because we moved their subjects when we migrated their
data to CGSpace.