roles/dspace: Set up sitemap redirect in nginx
We need to tell Google where to find the content that was living on the
CGIAR Library. They can stumble upon it and eventually map out the new
locations, but it's better if we submit the sitemap to the Google Search
console and let their bot figure it out in a structured way.