roles/dspace: Add nginx redirect for library.cgiar.org
We need this HTML file to be present to be able to add the CGIAR Library
domain to the Google Search Console. All other requests to this domain
should generally be passed to CGSpace. The Google Search Console allows
us to add sitemaps and control how Google indexes and displays results
to this domain.