roles/dspace: Add useragent mapping for bad client IPs
We want to be able to assign bots to a single Tomcat session using the
Crawler Session Manager Valve but some bots make thousands of requests
and don't even provide a user agent we can match on. Later versions of
Tomcat allow matching clients using IP but for now we'll just do it in
nginx.