roles/dspace: Explicitly set headers in nginx location block
Header directives are inherited by child blocks if and only if there are
no more headers being set in the current level. Because we're setting a
header for the user agent in this location block we need to explicitly
set the global headers again via including the proxy_params config. The
most important of these is the Host header, which will otherwise be set
to "tomcat_http" instead of the original $http_host that was requested
by the client.

See: https://nginx.org/en/docs/http/ngx_http_proxy_module.html#proxy_set_header