Install htop in compute role instead of common
This package comes from EPEL which isn't gauranteed to be installed on
all CentOS machines necessarily. We do have EPEL installed on compute
nodes, however, as GlusterFS relies on it, and to be honest those are
really our only CentOS machines anyways, so it makes sense to move it
here.