roles/common: Avoid using dynamic includes
Dynamic task includes make playbooks smaller and allow you to use some
neat tricks to include tasks based on a host's facts, but as of Ansible
2.4.0.0 this behavior seems to have changed. After hours of chasing my
tail around in circles trying to figure out why some of my tagged tasks
were not running, I noticed that changing these to static imports works
as expected.

It seems to me that dynamic includes are very complex and it is simpler
to rework playbooks to use static imports. This increases the size of
playbooks due to the repetition of includes, but the benefit is a much
easier to understand playbook flow.

See: https://docs.ansible.com/ansible/latest/playbooks_reuse.html