roles/compute: Add note about EPEL mirror
We can manage epel.repo eventually but I don't want to have to deal with
that right now, as I'm doing a lot of testing of CentOS 7 locally, where
I can't reach the HPC mirror quickly, not to mention we're thinking of a
move to HTTPS on hpc.ilri.cgiar.org. For now I need to focus on updating
these playbooks for migrating the cluster to CentOS 7.