roles/compute: Use ansible_host for host IP in /etc/hosts
The ansible_default_ipv4 variable comes from host facts, which may or
may not exist at the time of running the playbook. For example, I had
cleaned my local fact cache and during playbook execution facts were
only gathered for one machine I had limited, and therefore the facts
were not available for the others I was referencing in the groups.

This uses the ansible_host variable from host_vars instead, which is
probably always the same IP in effect, but could technically be diff-
erent.