Move /etc/hosts management from compute to common role
Storage and compute hosts rely quite a bit on being able to find each
other for yum repos, SLURM, LDAP, etc.

This task had previously been in the compute role but then the common
role was failing when it looked for HPC's CentOS mirror. Anyways, this
makes more sense to be in common rather than compute, because it will
also run on storage nodes.