group_vars/dspace: Bump PostgreSQL max_connections
This number is derived from the number of connections given to DSpace
in its db.maxconnections variable, which is currenly set to 90. I had
come up with a loose formula for estimating the appropriate limit for
PostgreSQL's max_connections based on the fact that every DSpace web
application can technically use up to db.connections *each*, meaning
that the system's PostgreSQL max_connections should at least be this
value times three, plus three reserved for the superuser.

90 * 3 + 3 = 273

Anyways, DSpace has been crashing a lot lately due to running out of
PostgreSQL connections, so we just need to increase this for now. I
would eventually like to set up a pool of connections so that ALL of
the DSpace web applications use one pool!