roles/compute: Remove note about SLURM head node
I don't want to worry about the SLURM head node for now, as we'd have to
start managing the config file with database passwords. For now, we only
need to create the SLURM user and install the common SLURM packages.

After all, it's not like we're provisioning lots of head nodes!