group_vars/compute: Use systemd automount for gluster volumes
Network mount points are failing because the network link is not *really* up
by the time the system tries to mount them. Adding a LINKDELAY used to work
but as of CentOS 7 the same hardware is having problems mounting the network
file systems, even with delays of 20, 40, and 60 seconds!

systemd's automount functionality allows volumes to be mounted on the fly as
they are accessed, but for some reason I had issues with the mount ordering
of volumes that are mounted below others (like data and mirror). The options
for controlling before and after dependencies don't seem to help. I decided
to move the video and mirror mounts to /srv, which is part of the root file
system, to work around this.

Also note that as of CentOS 7 the init system is now systemd, which already
knows that glusterfs is a network file system so we don't need to use the
_netdev option in /etc/fstab anymore.

See: http://codingberg.com/linux/systemd_when_to_use_netdev_mount_option