roles/compute: Configure SLURM to use limits
We set several limits on user processes in /etc/security/limits.d but
as of CentOS 7 and or SLURM 17.11 limits for user processes come from
the slurmd parent process. The slurm-pam_slurm package allows you to
configure SLURM to get limits from PAM.

For example, we want all user processes to be run at nice level of 19
instead of at 0.

See: https://hpcworks.wordpress.com/2017/05/29/setup-slurm-pam-plugin-on-centos-7/