roles/common: Run /etc/hosts task on HPC
This task is actually not destructive so it's pretty safe. Previously I
had not wanted to run it on HPC itself because there were a few other
hosts that HPC needs to talk to that aren't compute or storage servers,
and I couldn't think of a way to do that. But as this task is does not
do anything destructive it is safe to run there.