group_vars/dspace: Adjust random_page_cost
This tells the PostgreSQL query scheduler that accessing random pages
does not "cost" anything, as we're using a solid-state disk. This is
apparently a well-known tweak for running PostgreSQL on SSDs, and in
testing for a few weeks on DSpace Test this showed to greatly lessen
the number of connections to the database.