roles/dspace: Use a JNDI database pool for PostgreSQL
We have been having database pool issues for years, especially in the
last half of 2017 when the site grew in popularity with both bots and
other "real" users like our partners harvesting us. DSpace's database
stuff is pretty ghetto in that the default pool that is configured in
dspace.cfg is per web application, ie XMLUI, JSPUI, REST, etc all use
their own pool.

This setup is more or less according to the Tomcat JNDI documentation
but borrows some initial pool parameters from Mark Wood's comments on
the DSpace issue tracker[0]. I will monitor this and adjust it later.

The PostgreSQL JDBC library comes directly from the upstream site. It
is possible to use DSpace's bundled library by copying it to Tomcat's
library directory, but since we need to copy it anyways, I thought it
would be cleaner to get a fresh copy from upstream.

[0] https://jira.duraspace.org/browse/DS-3564
[1] https://jdbc.postgresql.org/download.html