roles/dspace: Increase nginx map hash bucket size
The default is apparently 64, but in order to map larger values like
user agents you need to increase it. According to the nginx docs the
value should be a multiple of the CPU's cache alignment size.

For us the cache alignment is 64:

  # cat /proc/cpuinfo | grep cache_alignment | head -n1
  cache_alignment : 64

See: https://nginx.org/en/docs/hash.html