roles/compute: Install firefox for Dedan
He had been using firefox as a module for years apparently, but it was
version 33 and the current stable version of Firefox is 58. I removed
the module because it was very, very old but decided to install yum's
version so it at least gets some updates. Currently that is version 52
enterprise support release (ESR).