roles/munin: Add jmx_dspace_sessions
This will allow us to monitor the number of current active sessions in
each DSpace application that supports it (ie, rest does not appear to
use sessions). The code is based on jmx_tomcat_dbpools that comes with
the munin-plugins-java Ubuntu package.

Requires enabling JMX monitoring in Tomcat's JAVA_OPTS.