roles/dspace: Increase JDBC removeAbandonedTimeout
The default is 60 seconds, so let's try it at 90. One user noticed that
they got a "socket closed" error in DSpace, and when I looked in the log
files there have been five or six of those errors in the past few days
since I enabled the timeout.

I will also enable logging of abandoned connections.