roles/dspace: Don't remove abandoned PostgreSQL connections
I've been running this for a few weeks and it seems to be helping with
the abandoned connections, but it might be too aggressive. One user is
now getting "Socket closed" errors on the DSpace submissions page very
often. Logically this is related, as he also says loading that page is
very slow, so I want to instead try to figure out why that page cannot
load within 60 or 90 or even 180 seconds and the query gets killed.