roles/common: Do not install heirloom-mailx on Ubuntu 18.04
This package has apparently been discontinued in favor of s-nail as
of Ubuntu 16.04(!) so we need to re-work our monitoring and backup
scripts to work with this.

It's likely that our scripts are not even working anyways because of
changes at the email hosts, not to mention the churn of admin users,
passwords, etc over the last two years.