roles/dspace: Install Oracle Java before Tomcat
The oracle-java8-installer package from the Oracle Java PPA provides
the default-jre-headless dependency required by Tomcat, yet actually
end up with both because we install Tomcat first. I never noticed it
until I was testing Ubuntu 18.04 because the system default Java was
always version 8, but now it is version 9.

Eventually we should switch to using Ubuntu's java8-runtime-headless
and stop depending on the third-party hackery in the Oracle Java PPA.