roles/dspace: Specifically install OpenJDK 8 on Ubuntu 18.04
Ubuntu 18.04 is going to be using Java 10 as the default[0]. This means
that packages like tomcat8-common that rely on on default-jre will now
pull Java 10 in unless we satisfy the JRE runtime requirement manually
by installing openjdk-8-jdk-headless ourselves. Unfortunately it seems
that there is a new problem with maven, as it seems to have a hard dep
that pulls in Java 10 still. This currently breaks our playbooks.

[0] https://lists.ubuntu.com/archives/ubuntu-release/2018-February/004275.html