roles/dspace:
Explicitly install OpenJDK 8 first in a separate transaction so that
tomcat8 doesn't pull in Java 10+ via its default-jre dependency.

Also, it seems there is a bug with the dependency resolution in apt
that causes maven to pull in openjdk-11-jre-headless to satisfy the
java7-runtime-headless dependency despite openjdk-8-jre-headless ac-
tually providing this. According to a user on #ubuntu+1 on Freenode
IRC the apt developer says it is a bug and it "can happen."