roles/munin: Explicitly install OpenJDK 8 JRE
Ubuntu 18.04 will be shipping Java 11 as the default-jre, so unless we
manually install another one first, munin-plugins-java will pull that
one in as a dependency.

This is not optimal, and Java in Ubuntu 18.04 is kinda turning out to
be a pain in the ass. Also, we are already doing this in the dspace
role as well, so this is really dirty and will need to be cleaned up
eventually.