roles/dspace: Tag 35.227.26.162 as a bot in nginx
It is making thousands of requests to XMLUI every day and creates a
new Tomcat session for each request. From looking at the logs I see
that it does not specify a user agent, so I need to tag it as a bot
so that Tomcat's Crawler Session Manager Valve recycles its session.