roles/dspace: Add Pcore-HTTP bot to Tomcat's Crawler Session Manager
This bot makes thousands of requests to dynamic search pages despite
them being forbidden in robots.txt which causes high load on the se-
rver. To its credit, the bot does seem to re-use its Tomcat sessions
and I am already rate limiting it in nginx, so adding it to the list
here is more of a backup measure.