roles/common: Set SSH's MaxAuthTries to 3
The SSH default for this is either 4 or 6 depending on the version,
but both of those seem to be too high. I think for now I'll set it
to 3. Eventually I'd like to find a way to ban bots using firewalld
or fail2ban that works on all our systems.