roles/common: Remove iptables-persistent on Ubuntu 15 and up
While upgrading a few systems from Ubuntu 14.04 to 16.04 I realized
that we need to remove the iptables-persistent service. This wasn't
a concern when we first deployed Ubuntu 16.04 two years ago, but we
need to do it for system idempotency. This could have led to errors
with the firewall that would be hard to diagnose!