roles/common: Cleanup old iptables in new Debian
We're using firewalld in Debian 8, but we need to cater for situations
where a Debian 7 system is being upgraded to Debian 8. In this case we
need to remove the iptables-peristent service, clean up its rules, and
restart firewalld if necessary to make sure the state of the firewall
is as intended.