roles/common: Make restic binary available
For now we simply download the statically linked binary from GitHub
because restic is only in Linux package repositories as of very re-
cently. I decided to download the binary and encode the version nu-
mber into the name so I didn't have to deal with parsing the versi-
on from the output of `restic version` and because I'm not sure ab-
out the restic versionioning policy and didn't want to have to make
some horribly unreadable regexes to cover tons of scenarios.