roles/dspace: Handle server service should use "forking"
After a recent reboot I noticed that the Handle server service had
been started by systemd but exited successfully shortly after. I'm
not sure if this is a bug or not, because as far as I can remember,
the service was working as I had written it several months ago.

Anyways, upon closer inspection, it does seem like there might be
an issue because the service is calling the standard DSpace Handle
server shell script, which launches with nohup and then backgrounds
the process. I'm not sure, but it seems like systemd might assume
that the service is completed with the parent process exits? In any
case, changing this to "forking" fixes the issue so that the service
stays up permanently.