roles/dspace: Add "most-popular" page to list of dynamic pages
We can't use wildcards to match these patterns in robots.txt, so we
need to use X-Robots-Tag header to inform compliant bots that they
should not index the page or follow links on it. These pages are not
useful to bots because they do not contain any actual content and
they cause a high load on the server.