roles/common: Allow disabling of timestamps in MySQL backups
We used to do a nightly MySQL backup and have a cron job clean up
backup files older than a week or two. Having the timestamp in the
file name made it easy to get a backup from a previous day, though
it required much more disk space.

Now, tools like restic can back up to offsite object storage like
Amazon S3 and rotate the backups according to some scheme, meaning
that we no longer need to use timestamps in the file name itself.
The restic backup repository contains the timestamps so it is easy
to access previous backups if need be.