roles/common: Add configurable restic backup service
This allows us to configure restic backups to Amazon S3 for each host
as well as to store the credentials here in the repository. Note that
the S3 bucket should already exist (create it manually with s3cmd).

Configure the restic password and S3 bucket path in host_vars:

    restic_password: nice_password
    restic_repository: s3:s3.amazonaws.com/hostname.ilri.org

By default only /etc is backed up, but you can override it by adding
something like this in host_vars:

    restic_backup_paths:
      - /etc
      - /home/backup/mysql