roles/dspace: Run Tomcat from tarball
We now run Tomcat from an upstream stable tarball instead of from the
operating system's tomcat7 package. This was born out of necessity to
be able to use Tomcat 7 on Ubuntu 18.04 because that release is using
Tomcat 8.5, which has problems with DSpace 5.x and 6.x currently.

The bonus of this is that we are no longer limited to any particular
operating system version and will be able to track upstream Tomcat
releases. Ironically this is similar to a container workflow.