roles/solr: Use PIDFile in solr.service
Solr's wrapper script doesn't seem to work well with systemd because
it forks a child process and systemd loses track of it. Solr already
writes its PID to a file at a known path so we can simply pass this
to systemd to let it know the real PID of the spawned process.

The systemd.service man page says this:

    Unless Type=forking is set, the process started via this command
    line will be considered the main process of the daemon.

See this discussion of a similar issue with systemd and the Tomcat
catalina.sh script, where the optimal solution is to use the script
to exec Tomcat rather than just run it in the background:

https://jdebp.eu/FGA/systemd-house-of-horror/tomcat.html