roles/dspace: Add new bot user agent to "badbots" nginx rate limiting
There are at least thirty IPs (all on Amazon) using this user agent
and making thousands of requests per day each, concurrently! This
causes a high load on the server and gives us nothing in return, as
they are generally indexing dynamic pages like Discovery that are
forbidden in robots.txt anyways!