roles/dspace: Add more validation to Tomcat JDBC pool
We are having problems with XMLUI pages being blank and the errors
in the DSpace log hint at some PostgreSQL connection issues:

    2019-03-22 21:21:34,356 ERROR org.dspace.storage.rdbms.DatabaseManager @ SQL QueryTable Error -
    org.postgresql.util.PSQLException: This statement has been closed.

I want to make sure we always test connections in the pool: before
using them, after returning them, when they are idle, etc. Also, I
will enable logging of validation failures and add two new JDBC in-
terceptors to keep track of connection state and close statements
when connections are returned to the pool.