roles/dspace: Add new bots to Tomcat and nginx config
Any client with the user agent "curl" should be marked as a bot by
Tomcat and forced to re-use their session variable. GuzzleHttp is
a PHP library for making HTTP requests and I've seen clients using
this library to hit dynamic pages on CGSpace so I will rate limit
them.