roles/common: Add task to remove old hosts from /etc/hosts
We *add* hosts to /etc/hosts files, but once a host dies and we re-
move it from the Ansible inventory there is no way to idempotently
reflect that. In this case we had a few storage servers die and we
need them to be removed, so let's just make a new task with those
hosts hard coded.