roles/compute: Add more packages
These are needed by some R libraries. It's a bit messy because, for
example, the R library poppr needs another library called sf, which
needs some system packages like gdal, geos, and proj (and of course
sf is not happy with CentOS 7's version of gdal or proj so I needed
to compile them manually).