roles/common: Adjust mail configuration for Ubuntu hosts
Only use heirloom-mailx on Ubuntu hosts before (and not including)
version 16.04. Otherwise, Ubuntu 16.04 and up have s-nail. We don't
need to test for Debian family because we're only using Ubuntu, so
it doesn't make sense to test the family just in case there might
be an actual Debian host (we switched to using only Ubuntu LTS a
few years ago).