roles/common: Add TCP 443 to default rules for web access
A few years ago HTTPS was less common so we only added TCP 80 to the
default "web" rules. Any hosts using HTTPS could add it to the extra
iptables rules in their host_vars. Now HTTPS is much more ubiquitous
so it makes sense that it would be part of the default rules.