roles/solr: Move become_user to apply
This seems to be a new paradigm that I missed. become_user has been
a syntax error for a few months and I've been avoiding dealing with
it, but this syntax seems to at least run without error. I have not
run the task on a host using solr yet.