group_vars/compute: Use hpc-compute tuned profile
I'm not sure if we are considered "HPC" or not, but it seems like
we should use this profile instead of latency-throughput, as that
aims to "Broadly applicable tuning that provides excellent perfor-
mance across a variety of common server workloads".

The hpc-compute profile appears to have been added in tuned 2.11.0.