roles/dspace: Pass mapped user agent to more location blocks
We need to make sure we are passing the user agents mapped from IPs
to all relevant location blocks that might need to know whether the
request is from a human or not. The REST API, for example, registers
requests as statistic hits, so if something is repeatadely harvesting
the API programmatically you are going to get highly inflated numbers
in your item statistics. We learned this the hard way when harvesting
via AReS and Drupal.

It does not appear that the OAI interface registers statistics hits
for its requests.