roles/dspace: Include global proxy_params in more nginx blocks
Global headers get reset when we modify headers in location blocks.
Make sure to re-include the global ones in this case so that we pa-
ss on important information to Tomcat (ie so that it can decided if
the request is a bot and log its correct IP in the Solr stats!).