roles/dspace: Actually allow UptimeRobot in nginx config
I was getting sporadic "down" alerts from UptimeRobot due to nginx
rate limiting requests. The nginx error log shows that the requests
were indeed being rate limited despite me thinking I had allowed it
by mapping the UptimeRobot user agent to 0. The problem is that we
actually need to map the value to an empty string to make nginx not
evaluate it!