roles/dspace: Use a random string for the cookie password
I don't think this password matters much, only that it is private
and that the application knows and uses it to encode cookies. If
we change the password we can invalidate all existing sessions, but
that's not such a big deal of course as sessions automatically time
out after thirty minutes anyways.