roles/dspace: Downgrade Tomcat to 7.0.99
Both CGSpace and DSpace Test seem to have an issue with a creeping
number of "idle in transaction" and "waiting for lock" connections
in PostgreSQL starting around late April. This causes issues during
item submission and harvesting. I had previously tried to downgrade
the PostgreSQL JDBC driver, but even when I revert to DSpace's much
older PostgreSQL driver (which breaks JDBC pooling) I have the same
issues. Not to mention, DSpace Test is currently on DSpace 6.3 and
PostgreSQL 10.13 and we see the issues there too!