roles/dspace: Add nginx rewrite for invalid REST URLs
DSpace has a bug that causes the retrieveLink URLs for bitstreams
to not have "/rest" in them. I patched it, but it seems some users
of the API had been adding "/rest" to the strings themselves which
causes the URL to be invalid.

See: https://jira.lyrasis.org/browse/DS-3193