roles/dspace: Add Turnitin to Tomcat's Crawler Session Manager valve
This is the alternate user-agent of Turnitin that does not have the
word "bot" in it so it slips through the cracks of our defenses. We
need to force this bot to use the same session as other crawlers so
it doesn't cause high load on the server by creating a new session
for each request.