roles/dspace: Explicitly allow Googlebot and Twitterbot
They were getting lumped in with all the other bots in the aggress-
ive rate limiting in nginx so they were seeing lots of HTTP 5xx.

In reality Google is annoying, but it respects robots.txt and does
not crawl haphazardly. This aggressive rate limiting was even caus-
ing it to get HTTP 5xx on the sitemap, which it is supposed to use
in order to crawl items directly.

Twitterbot is probably just trying to fetch page metadata or image
when a user shares a link on Twitter and is less than 100 requests
per day.