roles/compute: Remove CPU time limit
This very old setting in security/limits.conf turns out to have been
causing issues with long-running processes such as rsync. In my test
I modified the limit to be 1 minute for myself, then tried copying a
large file over rsync. The process was killed after 11 minutes. Once
I removed the CPU limit entirely I was able to copy the file.

For reference, the error was:

    rsync: [sender] write error: Broken pipe (32)
    rsync error: unexplained error (code 255) at io.c(823) [sender=v3.2.3]