roles/common: Update comment about Mellanox driver/utils
The repository configuration is for each point release of RHEL, for
example 8.1, 8.2, etc... so I'm not sure how this out-of-tree driver
will work when we automatically update to CentOS 8.x+1. It might be
better to stick with the vanilla mlx5_core driver that comes in the
CentOS kernel, and just install the utils from Mellanox's repo.