group_vars/compute: Add global scratch
This is a non-redundant volume on GlusterFS that can be used for a
global scratch space accessible from any host on the cluster. It
will not be as fast as the local /var/scratch of course, but will
be useful for pipelines and such running on multiple hosts that
need access to shared data.