roles/compute: Add support for NVIDIA Tesla V100 GPU
This uses SLURM's Generic Resources (Gres). There are various ways
to configure this according to slurm.conf and gres.conf man pages.
The cgroup.conf change ensures that only jobs that have requested
the GPU can access it.

Currently we only have a GPU on compute06 so we just hard code some
stuff in the tasks for that node.