roles/dspace: Install libpq-dev
We had been relying on psycopg2-binary for a long time because it
was easier, but this is apparently not the recommended way to use
psycopg2 in production. We can simply install the PostgreSQL dev
headers which will allow us to build psycopg2 from source.

See: https://www.psycopg.org/docs/install.html#quick-install