roles/dspace: Add acl to DSpace system dependencies
This works around an issue in recent Ansible (?) when you are using
file-based modules like template with an unprivileged become_user.
The error is incredibly misleading:

TypeError: can only concatenate list (not "tuple") to list fatal

When you turn on -vvv you see that it's unrelated to the module's
action, and actually just an Ansible implementation detail. If we
install `acl` to provide the `setfacl` command this works fine.

See: https://docs.ansible.com/ansible/latest/user_guide/become.html