roles/dspace: Use /dev/urandom for SecureRandom in Java
This is why Tomcat has been taking minutes to start. I assumed it
was due to DSpace 6 or heavy load or something, but I happened to
see this message in the Tomcat journal today:

tomcat7[30294]: WARNING: Creation of SecureRandom instance for session ID generation using [SHA1PRNG] took [111,733] milliseconds.

Upon investigation I found that Tomcat is using /dev/random which
is blocking, and causes startup to take several minutes. For our
purposes /dev/urandom is perfectly sane and allows Tomcat to start
in thirty seconds or so instead of three or four minutes.

See: https://www.2uo.de/myths-about-urandom/