misc_plays/cleanup_firewalld.yml: Add support for Ubuntu 18.04
In addition to making sure that iptables and firewalld are gone, we
need to make sure that virtual guests are running a kernel supporting
nftables (as the linux-kvm and linux-virtual kernels in Ubuntu 18.04
do not). We explicitly install the HWE LTS kernel and reboot the host
as part of the cleanup. This prepares the host for being able to run
the firewall tasks in the main playbooks.

We don't need to worry about bare-metal hosts running Ubuntu 18.04,
as the generic kernel supports nftables (though we may want to move
them all to the HWE LTS kernel eventually).