roles/common: re-work ipset tasks
We can do something more clever than using multiple tasks for the
static and dynamic sets by adding a force parameter to each set.
This way we have less code in the playbook and we only overwrite
ipsets that are static and stored in git.

I tested this on Ubuntu 16.04, Ubuntu 20.04, and CentOS 7.