roles/dspace: Block robot user agent in nginx
This user agent appears to be a bot pretending to be a human:

  Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)

It makes hundreds of thousands of requests from hundreds of IPs each
month, most of which are on Amazon's network (ASN 14618). I am adding
them to the list of "bad" bots so they get heavily rate limited.