roles/common: Disable ssh-rsa host key algorithm
The ssh-rsa host key algorithm uses SHA-1 hash algorithm by default,
which has been known to be unsafe in general for several years, but
in specific since 2020[1]. As of OpenSSH 8.2 (2021-02-14) ssh-rsa was
marked as deprecated, as there are several good alternatives that are
available. Most of our servers are new enough to support ED25519, so
let's just use that across the board.

See: https://www.openssh.com/txt/release-8.2

[1] "SHA-1 is a Shambles: First Chosen-Prefix Collision on SHA-1 and
    Application to the PGP Web of Trust" Leurent, G and Peyrin, T
    (2020) https://eprint.iacr.org/2020/014.pdf