roles/compute: re-work handlers for slurm
We need to restart all slurm daemons if any slurm configs change,
but some of the daemons only run on the head node. Also, because
handlers are run in the order they are defined (not the order they
are notified), we list slurmdbd first, slurmctld second, and then
slurmd.

This way I can just run like this when we have config updates:

$ ansible-playbook compute.yml -l '!hpc' -t slurm
$ ansible-playbook compute.yml -l 'hpc' -t slurm

Make sure the compute nodes are running the latest configs and have
their daemons restarted, then do the head node.